chore: 안 쓰는 코드 및 파일 삭제
